pyproject.toml
[project]
name = "datahub-mcp"
version = "0.1.0"
description = "A unified MCP server providing seamless access to MySQL databases and Google Workspace APIs"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.18.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.185.0",
]